component cleanups
