+++ demos/react-convex-todolist/src/app/views/todo-lists/TodoListsModalLayout.tsx
@@ -1,4 +1,4 @@
-import TodoEditModalRoute from './TodoEditModalRoute';
+import TodoEditModalRoute from '@/components/todo-lists/TodoEditModalRoute';
 import TodoListsPage from './page';
 
 /**
