--- demos/react-convex-todolist/src/app/views/todo-lists/listDetailsFormTypes.ts
@@ -1,7 +1,0 @@
-/** List metadata edited in the todo list modal (Formik); persisted automatically. */
-export type ListDetailsFormValues = {
-  name: string;
-  notes: string;
-  tags: string[];
-  priority: number;
-};
