+++ demos/react-convex-todolist/src/components/widgets/ArchivedTodoListsPanel.tsx
@@ -1,5 +1,5 @@
 import { TODO_LISTS_ROUTE } from '@/app/router';
-import { formatListTaskSummary, listPriorityCaption } from '@/app/views/todo-lists/listFormUtils';
+import { formatListTaskSummary, listPriorityCaption } from '@/components/todo-lists/listFormUtils';
 import type { TodoListWithCountsRow } from '@/library/powersync/AppSchema';
 import { List, Paper, Typography } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
